Rename directory and add exports
