+++ src/index.tsx
@@ -1,16 +1,23 @@
-export * from './hooks/bindings/useExecutorchModule';
-export * from './hooks/natural_language_processing/useLLM';
-export * from './constants/modelUrls';
-
 // hooks
 export * from './hooks/computer_vision/useClassification';
 export * from './hooks/computer_vision/useObjectDetection';
 export * from './hooks/computer_vision/useStyleTransfer';
+
+export * from './hooks/natural_language_processing/useLLM';
+
+export * from './hooks/general/useExecutorchModule';
 
 // modules
 export * from './modules/computer_vision/ClassificationModule';
 export * from './modules/computer_vision/ObjectDetectionModule';
 export * from './modules/computer_vision/StyleTransferModule';
 
+export * from './modules/natural_language_processing/LLMModule';
+
+export * from './modules/general/ExecutorchModule';
+
 // types
 export * from './types/object_detection';
+
+// constants
+export * from './constants/modelUrls';
